Fix a couple of warnings reported by MSVC

+++ src/game.c
@@ -224,8 +224,8 @@ void game_remove_player_penguin(Game* self, int idx, Coords coords) {
   assert(penguin != NULL);
   // The penguin pointer will be within the boundaries of the penguins array,
   // this is legal.
-  int penguin_idx = penguin - player->penguins;
-  for (int i = penguin_idx; i < player->penguins_count - 1; i++) {
+  size_t penguin_idx = penguin - player->penguins;
+  for (size_t i = penguin_idx; i < player->penguins_count - 1; i++) {
     player->penguins[i] = player->penguins[i + 1];
   }
   player->penguins_count -= 1;
